FIX circular dependencies
