+++ lib/theming/src/themes/light-vars.ts
@@ -1,0 +1,33 @@
+import { color, typography, background } from '../base';
+
+export const themeVars = {
+  // Storybook-specific color palette
+  colorPrimary: '#FF4785', // coral
+  colorSecondary: '#1EA7FD', // ocean
+
+  // UI
+  appBg: background.app,
+  appContentBg: color.lightest,
+  appBorderColor: color.border,
+  appBorderRadius: 4,
+
+  // Fonts
+  fontBase: typography.fonts.base,
+  fontCode: typography.fonts.mono,
+
+  // Text colors
+  textColor: color.darkest,
+  textInverseColor: color.lightest,
+
+  // Toolbar default and active colors
+  barTextColor: color.mediumdark,
+  barSelectedColor: color.secondary,
+  barBg: color.lightest,
+
+  // Form colors
+  inputBg: color.lightest,
+  inputBorder: color.border,
+  inputTextColor: color.darkest,
+  inputBorderRadius: 4,
+};
+
