+++ scripts/compile-js.js
@@ -34,7 +34,9 @@ function babelify(options = {}) {
   const { watch = false, silent = true, errorCallback } = options;
 
   if (!fs.existsSync('src')) {
-    if (!silent) console.log('No src dir');
+    if (!silent) {
+      console.log('No src dir');
+    }
     return;
   }
